Add files via upload
This scripts aims to detect open reading frames (ORFs) in DNA sequences. If the user inserts an RNA sequence the program can't proceed and will notify the user. Sequences are inserted from the keyboard one at a time and are scanned for start codons. After detection of a start codon, the program scans for a stop codon in triplet-pacing (scaning of three residues at a time). One the codons are detected the user is notified about their positions. The sequence between a start and stop codon is considered as an ORF. The potential ORFs are subjected to "quality control" in order to eliminate ORFs that consist of only a start and stop codon -these sequences are not considered to be real ORFs. Once the scanning of the sequence is completed the user is asked if they would like to re-insert their sequence in order for the program to detect ORFs in the complementary strand. If yes, the complementary sequence is found and reversed and scanned for start and stop codons at the appropriate pacing. If any of the sequences contain no start and/or stop codons the user is notified.